cfg80211: Use new wiphy flag WIPHY_FLAG_DFS_OFFLOAD
When flag WIPHY_FLAG_DFS_OFFLOAD is defined, the driver would handle
all the DFS related operations. Therefore the kernel needs to ignore
the DFS state that it uses to block the userspace calls to the driver
through cfg80211 APIs. Also it should treat the userspace calls to
start radar detection as a no-op.

Please note that changes in util.c is not picked up explicitly.
Kernel v4.4.15 uses wrapper cfg80211_get_chans_dfs_required which takes
care of this change.

IRs-Fixed: 202686
Change-Id: I9dd2076945581ca67e54dfc96dd3dbc526c6f0a2